Refactor scores
