+++ bayesflow/scores/__init__.py
@@ -1,14 +1,10 @@
 """Scoring rules for point estimation."""
 
-from .scores import (
-    ScoringRule,
-    ParametricDistributionRule,
-    NormedDifferenceScore,
-    MedianScore,
-    MeanScore,
-    QuantileScore,
-    MultivariateNormalScore,
-)
+from .normed_difference_score import NormedDifferenceScore
+from .mean_score import MeanScore
+from .median_score import MedianScore
+from .quantile_score import QuantileScore
+from .multivariate_normal_score import MultivariateNormalScore
 
 from ..utils._docs import _add_imports_to_all
